Fix errors with recent compilers
OpenBSD decided to kill strncpy() due to it being consistently unsafe.
While it could be used safely, more often than not it wasn't.  All of
these invocations of strncpy() are hazardous, and many could well cause
security holes.

The recent compiler writers are correct to check use of strncpy(), but
we really should simply stick a fork in it.